admin_panel/requirements.txt
fastapi[standard]
uvicorn[standard]
jinja2
python-multipart
pandas
openpyxl
python-dotenv
sqlalchemy
asyncpg